refactor(ios): redesign genre selection as search-driven chips instead of full list
Selected genres now display as removable pill chips, with search results
capped at 10 shown only while typing. Extracts FlowLayout to a shared
component reused by NowPlayingDetailsCard.
